Rename derive_partialeq_or_partialord
