+++ src/ir/analysis/mod.rs
@@ -54,8 +54,8 @@ mod has_type_param_in_array;
 pub use self::has_type_param_in_array::HasTypeParameterInArray;
 mod derive_hash;
 pub use self::derive_hash::CannotDeriveHash;
-mod derive_partial_eq_or_partial_ord;
-pub use self::derive_partial_eq_or_partial_ord::CannotDerivePartialEqOrPartialOrd;
+mod derive_partialeq_or_partialord;
+pub use self::derive_partialeq_or_partialord::CannotDerivePartialEqOrPartialOrd;
 mod has_float;
 pub use self::has_float::HasFloat;
 mod sizedness;
